pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "langchain-zhipuai>=0.0.1",
    "langgraph>=1.0.7",
    "mcp>=1.26.0",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "zhipuai>=1.0.7",
]
